pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pytest-essentials"
version = "0.1.0"
authors = [
    { name="Pragadeshwar Vishnu", email="mkpvishnu@gmail.com" },
]
description = "A Pytest plugin providing essential utilities like soft assertions."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "pytest>=7.0",
]

[project.optional-dependencies]
allure = ["allure-pytest>=2.9.45"] # For Allure reporting integration

[project.entry-points."pytest11"]
pytest_essentials = "pytest_essentials.plugin"

[project.urls]
Homepage = "https://github.com/TalkTrove/pytest-essentials"
Repository = "https://github.com/TalkTrove/pytest-essentials"

[tool.setuptools.packages.find]
where = ["."]
include = ["pytest_essentials*"]
namespaces = false

[tool.pytest.ini_options]
markers = [
    "soft_assert_level(level): Mark test to change soft assertion failure level (broken, failed, passed).",
]